Cleanup VisualStyles enums
