+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/ColorProperty.cs
@@ -1,0 +1,61 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    [SuppressMessage("Microsoft.Design", "CA1008:EnumsShouldHaveZeroValue", Justification = "Maps to native enum.")]
+    public enum ColorProperty
+    {
+    	BorderColor = 3801,
+    	FillColor = 3802,
+    	TextColor = 3803,
+    	EdgeLightColor = 3804,
+    	EdgeHighlightColor = 3805,
+    	EdgeShadowColor = 3806,
+    	EdgeDarkShadowColor = 3807,
+    	EdgeFillColor = 3808,
+    	TransparentColor = 3809,
+    	GradientColor1 = 3810,
+    	GradientColor2 = 3811,
+    	GradientColor3 = 3812,
+    	GradientColor4 = 3813,
+    	GradientColor5 = 3814,
+    	ShadowColor = 3815,
+    	GlowColor = 3816,
+    	TextBorderColor = 3817,
+    	TextShadowColor = 3818,
+    	GlyphTextColor = 3819,
+    	GlyphTransparentColor = 3820,
+    	FillColorHint = 3821,
+    	BorderColorHint = 3822,
+    	AccentColorHint = 3823
+    //		TM_PROP(3801, TMT, BORDERCOLOR,      COLOR)       // color of borders for BorderFill
+    //		TM_PROP(3802, TMT, FILLCOLOR,        COLOR)       // color of bg fill
+    //		TM_PROP(3803, TMT, TEXTCOLOR,        COLOR)       // color text is drawn in
+    //		TM_PROP(3804, TMT, EDGELIGHTCOLOR,     COLOR)     // edge color
+    //		TM_PROP(3805, TMT, EDGEHIGHLIGHTCOLOR, COLOR)     // edge color
+    //		TM_PROP(3806, TMT, EDGESHADOWCOLOR,    COLOR)     // edge color
+    //		TM_PROP(3807, TMT, EDGEDKSHADOWCOLOR,  COLOR)     // edge color
+    //		TM_PROP(3808, TMT, EDGEFILLCOLOR,  COLOR)         // edge color
+    //		TM_PROP(3809, TMT, TRANSPARENTCOLOR, COLOR)       // color of pixels that are treated as transparent (not drawn)
+    //		TM_PROP(3810, TMT, GRADIENTCOLOR1,   COLOR)       // first color in gradient
+    //		TM_PROP(3811, TMT, GRADIENTCOLOR2,   COLOR)       // second color in gradient
+    //		TM_PROP(3812, TMT, GRADIENTCOLOR3,   COLOR)       // third color in gradient
+    //		TM_PROP(3813, TMT, GRADIENTCOLOR4,   COLOR)       // forth color in gradient
+    //		TM_PROP(3814, TMT, GRADIENTCOLOR5,   COLOR)       // fifth color in gradient
+    //		TM_PROP(3815, TMT, SHADOWCOLOR,      COLOR)       // color of text shadow
+    //		TM_PROP(3816, TMT, GLOWCOLOR,        COLOR)       // color of glow produced by DrawThemeIcon
+    //		TM_PROP(3817, TMT, TEXTBORDERCOLOR,  COLOR)       // color of text border
+    //		TM_PROP(3818, TMT, TEXTSHADOWCOLOR,  COLOR)       // color of text shadow
+    //		TM_PROP(3819, TMT, GLYPHTEXTCOLOR,        COLOR)  // color that font-based glyph is drawn with
+    //		TM_PROP(3820, TMT, GLYPHTRANSPARENTCOLOR, COLOR)  // color of transparent pixels in GlyphImageFile
+    //		TM_PROP(3821, TMT, FILLCOLORHINT, COLOR)          // hint about fill color used (for custom controls)
+    //		TM_PROP(3822, TMT, BORDERCOLORHINT, COLOR)        // hint about border color used (for custom controls)
+    //		TM_PROP(3823, TMT, ACCENTCOLORHINT, COLOR)        // hint about accent color used (for custom controls)
+    }
+}
