+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/HorizontalAlign.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    public enum HorizontalAlign
+    {
+    	Left = 0,
+    	Center = 1,
+    	Right = 2,
+    //		TM_ENUM(0, HA, LEFT)
+    //		TM_ENUM(1, HA, CENTER)
+    //		TM_ENUM(2, HA, RIGHT)
+    }
+}
