+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/IconEffect.cs
@@ -1,0 +1,24 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    public enum IconEffect
+    {
+    	None = 0,
+    	Glow = 1,
+    	Shadow = 2,
+    	Pulse = 3,
+    	Alpha = 4,
+    //		TM_ENUM(0, ICE, NONE)
+    //		TM_ENUM(1, ICE, GLOW)
+    //		TM_ENUM(2, ICE, SHADOW)
+    //		TM_ENUM(3, ICE, PULSE)
+    //		TM_ENUM(4, ICE, ALPHA)
+    }
+}
