+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/ImageOrientation.cs
@@ -1,0 +1,18 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    public enum ImageOrientation 
+    {
+    	Vertical = 0,
+    	Horizontal = 1,
+    //		TM_ENUM(0, IL, VERTICAL)
+    //		TM_ENUM(1, IL, HORIZONTAL)
+    }
+}
