+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/ImageSelectType.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    public enum ImageSelectType
+    {
+    	None = 0,
+    	Size = 1,
+    	Dpi = 2,
+    //		TM_ENUM(0, IST, NONE)
+    //		TM_ENUM(1, IST, SIZE)
+    //		TM_ENUM(2, IST, DPI)
+    }
+}
