+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/IntegerProperty.cs
@@ -1,0 +1,63 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    [SuppressMessage("Microsoft.Design", "CA1008:EnumsShouldHaveZeroValue", Justification = "Maps to native enum.")]
+    public enum IntegerProperty
+    {
+    	ImageCount = 2401,
+    	AlphaLevel = 2402,
+    	BorderSize = 2403,
+    	RoundCornerWidth = 2404,
+    	RoundCornerHeight = 2405,
+    	GradientRatio1 = 2406,
+    	GradientRatio2 = 2407,
+    	GradientRatio3 = 2408,
+    	GradientRatio4 = 2409,
+    	GradientRatio5 = 2410,
+    	ProgressChunkSize = 2411,
+    	ProgressSpaceSize = 2412,
+    	Saturation = 2413,
+    	TextBorderSize = 2414,
+    	AlphaThreshold = 2415,
+    	Width = 2416,
+    	Height = 2417,
+    	GlyphIndex = 2418,
+    	TrueSizeStretchMark = 2419,
+    	MinDpi1 = 2420,
+    	MinDpi2 = 2421,
+    	MinDpi3 = 2422,
+    	MinDpi4 = 2423,
+    	MinDpi5 = 2424
+    //		TM_PROP(2401, TMT, IMAGECOUNT,        INT)    // the number of state images in an imagefile
+    //		TM_PROP(2402, TMT, ALPHALEVEL,        INT)    // (0-255) alpha value for an icon (DrawThemeIcon part)
+    //		TM_PROP(2403, TMT, BORDERSIZE,        INT)    // the size of the border line for bgtype=BorderFill
+    //		TM_PROP(2404, TMT, ROUNDCORNERWIDTH,  INT)    // (0-100) % of roundness for rounded rects
+    //		TM_PROP(2405, TMT, ROUNDCORNERHEIGHT, INT)    // (0-100) % of roundness for rounded rects
+    //		TM_PROP(2406, TMT, GRADIENTRATIO1,    INT)    // (0-255) - amt of gradient color 1 to use (all must total=255)
+    //		TM_PROP(2407, TMT, GRADIENTRATIO2,    INT)    // (0-255) - amt of gradient color 2 to use (all must total=255)
+    //		TM_PROP(2408, TMT, GRADIENTRATIO3,    INT)    // (0-255) - amt of gradient color 3 to use (all must total=255)
+    //		TM_PROP(2409, TMT, GRADIENTRATIO4,    INT)    // (0-255) - amt of gradient color 4 to use (all must total=255)
+    //		TM_PROP(2410, TMT, GRADIENTRATIO5,    INT)    // (0-255) - amt of gradient color 5 to use (all must total=255)
+    //		TM_PROP(2411, TMT, PROGRESSCHUNKSIZE, INT)    // size of progress control chunks
+    //		TM_PROP(2412, TMT, PROGRESSSPACESIZE, INT)    // size of progress control spaces
+    //		TM_PROP(2413, TMT, SATURATION,        INT)    // (0-255) amt of saturation for DrawThemeIcon() part
+    //		TM_PROP(2414, TMT, TEXTBORDERSIZE,    INT)    // size of border around text chars
+    //		TM_PROP(2415, TMT, ALPHATHRESHOLD,    INT)    // (0-255) the min. alpha value of a pixel that is solid
+    //		TM_PROP(2416, TMT, WIDTH,             SIZE)   // custom window prop: size of part (min. window)
+    //		TM_PROP(2417, TMT, HEIGHT,            SIZE)   // custom window prop: size of part (min. window)
+    //		TM_PROP(2418, TMT, GLYPHINDEX,        INT)    // for font-based glyphs, the char index into the font
+    //		TM_PROP(2419, TMT, TRUESIZESTRETCHMARK, INT)  // stretch TrueSize image when target exceeds source by this percent
+    //		TM_PROP(2420, TMT, MINDPI1,         INT)      // min DPI ImageFile1 was designed for
+    //		TM_PROP(2421, TMT, MINDPI2,         INT)      // min DPI ImageFile1 was designed for
+    //		TM_PROP(2422, TMT, MINDPI3,         INT)      // min DPI ImageFile1 was designed for
+    //		TM_PROP(2423, TMT, MINDPI4,         INT)      // min DPI ImageFile1 was designed for
+    //		TM_PROP(2424, TMT, MINDPI5,         INT)      // min DPI ImageFile1 was designed for
+    }
+}
