+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/OffsetType.cs
@@ -1,0 +1,42 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    public enum OffsetType
+    {
+    	TopLeft = 0,
+    	TopRight = 1,
+    	TopMiddle = 2,
+    	BottomLeft = 3,
+    	BottomRight = 4,
+    	BottomMiddle = 5,
+    	MiddleLeft = 6,
+    	MiddleRight = 7,
+    	LeftOfCaption = 8,
+    	RightOfCaption = 9,
+    	LeftOfLastButton = 10,
+    	RightOfLastButton = 11,
+    	AboveLastButton = 12,
+    	BelowLastButton = 13,
+    //		TM_ENUM(0, OT, TOPLEFT)
+    //		TM_ENUM(1, OT, TOPRIGHT)
+    //		TM_ENUM(2, OT, TOPMIDDLE)
+    //		TM_ENUM(3, OT, BOTTOMLEFT)
+    //		TM_ENUM(4, OT, BOTTOMRIGHT)
+    //		TM_ENUM(5, OT, BOTTOMMIDDLE)
+    //		TM_ENUM(6, OT, MIDDLELEFT)
+    //		TM_ENUM(7, OT, MIDDLERIGHT)
+    //		TM_ENUM(8, OT, LEFTOFCAPTION)
+    //		TM_ENUM(9, OT, RIGHTOFCAPTION)
+    //		TM_ENUM(10, OT, LEFTOFLASTBUTTON)
+    //		TM_ENUM(11, OT, RIGHTOFLASTBUTTON)
+    //		TM_ENUM(12, OT, ABOVELASTBUTTON)
+    //		TM_ENUM(13, OT, BELOWLASTBUTTON)
+    }
+}
