+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/TextMetricsCharacterSet.cs
@@ -1,0 +1,58 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    public enum TextMetricsCharacterSet
+    {
+    	Ansi = 0,
+    	Baltic = 186,
+    	ChineseBig5 = 136,
+    	Default = 1,
+    	EastEurope = 238,
+    	Gb2312 = 134,
+    	Greek = 161,
+    	Hangul = 129,
+    	Mac = 77,
+    	Oem = 255,
+    	Russian = 204,
+    	ShiftJis = 128,
+    	Symbol = 2,
+    	Turkish = 162,
+    	Vietnamese = 163,
+    	Johab = 130,
+    	Arabic = 178,
+    	Hebrew = 177,
+    	Thai = 222,
+    	//		#define ANSI_CHARSET            0
+    	//		#define BALTIC_CHARSET          186
+    	//		#define CHINESEBIG5_CHARSET     136
+    	//		#define DEFAULT_CHARSET         1
+    	//		#define EASTEUROPE_CHARSET      238
+    	//		#define GB2312_CHARSET          134
+    	//		#define GREEK_CHARSET           161
+    	//		#define HANGUL_CHARSET          129
+    	//		#define MAC_CHARSET             77
+    	//		#define OEM_CHARSET             255
+    	//		#define RUSSIAN_CHARSET         204
+    	//		#define SHIFTJIS_CHARSET        128
+    	//		#define SYMBOL_CHARSET          2
+    	//		#define TURKISH_CHARSET         162
+    	//		#define VIETNAMESE_CHARSET      163
+
+    	// Korean
+    	//		#define JOHAB_CHARSET           130
+
+    	// Middle East
+    	//		#define ARABIC_CHARSET          178
+    	//		#define HEBREW_CHARSET          177
+
+    	// Thai
+    	//		#define THAI_CHARSET            222
+    }
+}
