+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/ToolBarState.cs
@@ -1,0 +1,21 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    [SuppressMessage("Microsoft.Design", "CA1008:EnumsShouldHaveZeroValue")]
+    public enum ToolBarState
+    {
+        Normal = 1,
+        Hot = 2,
+        Pressed = 3,
+        Disabled = 4,
+        Checked = 5,
+        HotChecked = 6
+    }
+}
