+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/TrackBarThumbState.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    [SuppressMessage("Microsoft.Design", "CA1008:EnumsShouldHaveZeroValue")]
+    public enum TrackBarThumbState
+    {
+        Normal = 1,
+        Hot = 2,
+        Pressed = 3,
+        //Focused = 4, 
+        Disabled = 5
+    }
+}
