+++ src/System.Windows.Forms/src/System/Windows/Forms/VisualStyles/TrueSizeScalingType.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.MSInternal", "CA905:SystemAndMicrosoftNamespacesRequireApproval", Scope = "namespace", Target = "System.Windows.Forms.VisualStyles")]
+
+namespace System.Windows.Forms.VisualStyles
+{
+    public enum TrueSizeScalingType
+    {
+    	None = 0,
+    	Size = 1,
+    	Dpi = 2,
+    //		TM_ENUM(0, TSST, NONE)
+    //		TM_ENUM(1, TSST, SIZE)
+    //		TM_ENUM(2, TSST, DPI)
+    }
+}
